// website/src/modules/donation-certificates/donation-certificate.types.ts
export type DonationCertificateTableViewRow = {
	id: string;
	year: number;
	contributorFirstName: string;
	contributorLastName: string;
	email: string;
	storagePath: string;
	createdAt: Date;
};

export type DonationCertificateTableQuery = {
	page: number;
	pageSize: number;
	search: string;
	sortBy?: string;
	sortDirection?: 'asc' | 'desc';
};

export type DonationCertificatePaginatedTableView = {
	tableRows: DonationCertificateTableViewRow[];
	totalCount: number;
};

export type YourDonationCertificateTableViewRow = {
	id: string;
	year: number;
	language: string | null;
	storagePath: string | null;
	createdAt: Date;
};

export type YourDonationCertificateTableQuery = {
	page: number;
	pageSize: number;
	search: string;
	sortBy?: string;
	sortDirection?: 'asc' | 'desc';
};

export type YourDonationCertificatePaginatedTableView = {
	tableRows: YourDonationCertificateTableViewRow[];
	totalCount: number;
};

export type DonationCertificateError = 'technicalError' | 'bucketMissing' | 'alreadyExists' | 'noContributions';